requirements.txt
# Core Dagster
dagster==1.6.0
dagster-webserver==1.6.0
dagster-postgres==0.22.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.19

# Utilities
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0